+++ sdk/databoxedge/azure-mgmt-databoxedge/azure/mgmt/databoxedge/v2019_07_01/_data_box_edge_management_client.py
@@ -16,6 +16,7 @@ if TYPE_CHECKING:
     from typing import Any, Optional
 
     from azure.core.credentials import TokenCredential
+    from azure.core.pipeline.transport import HttpRequest, HttpResponse
 
 from ._configuration import DataBoxEdgeManagementClientConfiguration
 from .operations import Operations
@@ -38,31 +39,31 @@ class DataBoxEdgeManagementClient(object):
     """The DataBoxEdge Client.
 
     :ivar operations: Operations operations
-    :vartype operations: azure.mgmt.databoxedge.operations.Operations
+    :vartype operations: azure.mgmt.databoxedge.v2019_07_01.operations.Operations
     :ivar devices: DevicesOperations operations
-    :vartype devices: azure.mgmt.databoxedge.operations.DevicesOperations
+    :vartype devices: azure.mgmt.databoxedge.v2019_07_01.operations.DevicesOperations
     :ivar alerts: AlertsOperations operations
-    :vartype alerts: azure.mgmt.databoxedge.operations.AlertsOperations
+    :vartype alerts: azure.mgmt.databoxedge.v2019_07_01.operations.AlertsOperations
     :ivar bandwidth_schedules: BandwidthSchedulesOperations operations
-    :vartype bandwidth_schedules: azure.mgmt.databoxedge.operations.BandwidthSchedulesOperations
+    :vartype bandwidth_schedules: azure.mgmt.databoxedge.v2019_07_01.operations.BandwidthSchedulesOperations
     :ivar jobs: JobsOperations operations
-    :vartype jobs: azure.mgmt.databoxedge.operations.JobsOperations
+    :vartype jobs: azure.mgmt.databoxedge.v2019_07_01.operations.JobsOperations
     :ivar nodes: NodesOperations operations
-    :vartype nodes: azure.mgmt.databoxedge.operations.NodesOperations
+    :vartype nodes: azure.mgmt.databoxedge.v2019_07_01.operations.NodesOperations
     :ivar operations_status: OperationsStatusOperations operations
-    :vartype operations_status: azure.mgmt.databoxedge.operations.OperationsStatusOperations
+    :vartype operations_status: azure.mgmt.databoxedge.v2019_07_01.operations.OperationsStatusOperations
     :ivar orders: OrdersOperations operations
-    :vartype orders: azure.mgmt.databoxedge.operations.OrdersOperations
+    :vartype orders: azure.mgmt.databoxedge.v2019_07_01.operations.OrdersOperations
     :ivar roles: RolesOperations operations
-    :vartype roles: azure.mgmt.databoxedge.operations.RolesOperations
+    :vartype roles: azure.mgmt.databoxedge.v2019_07_01.operations.RolesOperations
     :ivar shares: SharesOperations operations
-    :vartype shares: azure.mgmt.databoxedge.operations.SharesOperations
+    :vartype shares: azure.mgmt.databoxedge.v2019_07_01.operations.SharesOperations
     :ivar storage_account_credentials: StorageAccountCredentialsOperations operations
-    :vartype storage_account_credentials: azure.mgmt.databoxedge.operations.StorageAccountCredentialsOperations
+    :vartype storage_account_credentials: azure.mgmt.databoxedge.v2019_07_01.operations.StorageAccountCredentialsOperations
     :ivar triggers: TriggersOperations operations
-    :vartype triggers: azure.mgmt.databoxedge.operations.TriggersOperations
+    :vartype triggers: azure.mgmt.databoxedge.v2019_07_01.operations.TriggersOperations
     :ivar users: UsersOperations operations
-    :vartype users: azure.mgmt.databoxedge.operations.UsersOperations
+    :vartype users: azure.mgmt.databoxedge.v2019_07_01.operations.UsersOperations
     :param credential: Credential needed for the client to connect to Azure.
     :type credential: ~azure.core.credentials.TokenCredential
     :param subscription_id: The subscription ID.
@@ -115,6 +116,24 @@ class DataBoxEdgeManagementClient(object):
             self._client, self._config, self._serialize, self._deserialize)
         self.users = UsersOperations(
             self._client, self._config, self._serialize, self._deserialize)
+
+    def _send_request(self, http_request, **kwargs):
+        # type: (HttpRequest, Any) -> HttpResponse
+        """Runs the network request through the client's chained policies.
+
+        :param http_request: The network request you want to make. Required.
+        :type http_request: ~azure.core.pipeline.transport.HttpRequest
+        :keyword bool stream: Whether the response payload will be streamed. Defaults to True.
+        :return: The response of your network call. Does not do error handling on your response.
+        :rtype: ~azure.core.pipeline.transport.HttpResponse
+        """
+        path_format_arguments = {
+            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
+        }
+        http_request.url = self._client.format_url(http_request.url, **path_format_arguments)
+        stream = kwargs.pop("stream", True)
+        pipeline_response = self._client._pipeline.run(http_request, stream=stream, **kwargs)
+        return pipeline_response.http_response
 
     def close(self):
         # type: () -> None