+++ sdk/kusto/azure-mgmt-kusto/azure/mgmt/kusto/models/__init__.py
@@ -52,14 +52,20 @@ try:
     from ._models_py3 import Operation
     from ._models_py3 import OperationDisplay
     from ._models_py3 import OperationListResult
+    from ._models_py3 import OperationResult
     from ._models_py3 import OptimizedAutoscale
     from ._models_py3 import ProxyResource
     from ._models_py3 import ReadOnlyFollowingDatabase
     from ._models_py3 import ReadWriteDatabase
     from ._models_py3 import Resource
+    from ._models_py3 import Script
+    from ._models_py3 import ScriptCheckNameRequest
+    from ._models_py3 import ScriptListResult
     from ._models_py3 import SkuDescription
     from ._models_py3 import SkuDescriptionList
     from ._models_py3 import SkuLocationInfoItem
+    from ._models_py3 import SystemData
+    from ._models_py3 import TableLevelSharingProperties
     from ._models_py3 import TrackedResource
     from ._models_py3 import TrustedExternalTenant
     from ._models_py3 import VirtualNetworkConfiguration
@@ -109,14 +115,20 @@ except (SyntaxError, ImportError):
     from ._models import Operation  # type: ignore
     from ._models import OperationDisplay  # type: ignore
     from ._models import OperationListResult  # type: ignore
+    from ._models import OperationResult  # type: ignore
     from ._models import OptimizedAutoscale  # type: ignore
     from ._models import ProxyResource  # type: ignore
     from ._models import ReadOnlyFollowingDatabase  # type: ignore
     from ._models import ReadWriteDatabase  # type: ignore
     from ._models import Resource  # type: ignore
+    from ._models import Script  # type: ignore
+    from ._models import ScriptCheckNameRequest  # type: ignore
+    from ._models import ScriptListResult  # type: ignore
     from ._models import SkuDescription  # type: ignore
     from ._models import SkuDescriptionList  # type: ignore
     from ._models import SkuLocationInfoItem  # type: ignore
+    from ._models import SystemData  # type: ignore
+    from ._models import TableLevelSharingProperties  # type: ignore
     from ._models import TrackedResource  # type: ignore
     from ._models import TrustedExternalTenant  # type: ignore
     from ._models import VirtualNetworkConfiguration  # type: ignore
@@ -128,6 +140,7 @@ from ._kusto_management_client_enums import (
     BlobStorageEventType,
     ClusterPrincipalRole,
     Compression,
+    CreatedByType,
     DataConnectionKind,
     DatabasePrincipalRole,
     DatabasePrincipalType,
@@ -144,6 +157,7 @@ from ._kusto_management_client_enums import (
     ProvisioningState,
     Reason,
     State,
+    Status,
     Type,
 )
 
@@ -193,14 +207,20 @@ __all__ = [
     'Operation',
     'OperationDisplay',
     'OperationListResult',
+    'OperationResult',
     'OptimizedAutoscale',
     'ProxyResource',
     'ReadOnlyFollowingDatabase',
     'ReadWriteDatabase',
     'Resource',
+    'Script',
+    'ScriptCheckNameRequest',
+    'ScriptListResult',
     'SkuDescription',
     'SkuDescriptionList',
     'SkuLocationInfoItem',
+    'SystemData',
+    'TableLevelSharingProperties',
     'TrackedResource',
     'TrustedExternalTenant',
     'VirtualNetworkConfiguration',
@@ -210,6 +230,7 @@ __all__ = [
     'BlobStorageEventType',
     'ClusterPrincipalRole',
     'Compression',
+    'CreatedByType',
     'DataConnectionKind',
     'DatabasePrincipalRole',
     'DatabasePrincipalType',
@@ -226,5 +247,6 @@ __all__ = [
     'ProvisioningState',
     'Reason',
     'State',
+    'Status',
     'Type',
 ]